minor refactor

+++ src/main/java/me/steffenjacobs/syntacticjsonmatcher/service/matcher/KeyMatcher.java
@@ -140,7 +140,7 @@ public class KeyMatcher {
 					max = entr.getValue();
 				}
 			}
-			return PENALTY_SYNONYM * (synonymsSource.get(target) / max);
+			return PENALTY_SYNONYM * (synonymsSource.get(target) / (double)max);
 		}
 		return 0;
 	}
